Amortize the next album fetching time
It fetches an new album on start up and stores it in memory. Then, in a
new concurrent task, it goes and fetches the next album after that.
There's a _bit_ of wonkiness with the current setup if there's a need
for a restart in between things, it'll need to manually update the user.
I think this can probably be solved by breaking out the user update with
the next album fetch as an async process after the fact, but I only have
so much time, so it'll be fine for now and I can ajust manually if the
need arises